+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -20,6 +20,8 @@
   per `cardano-ledger-binary-1.4.0.0`
 * Add a function `expectRawEqual`
 * Add `assertColorFailure`
+* Add `Test.Cardano.Ledger.Core.Binary.CDDL` containing core CDDL definitions
+  shared across eras.
 
 ## 1.14.0.0